ARQ-387 Moved MockDrone to its own test subpackage
